backend/requirements.txt
# --- API & Server ---
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# --- Database ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# --- AI & Data ---
# Core logic used by your incident_classifier.pkl
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2